+++ src/hkimjp/konpy2/comments.clj
@@ -1,11 +1,11 @@
 (ns hkimjp.konpy2.comments
   (:require
+   [nextjournal.markdown :as md]
    [taoensso.telemere :as t]
-   ; [hkimjp.carmine :as c]
    [hkimjp.datascript :as ds]
    [hkimjp.konpy2.response :refer [hx redirect page]]
-   [hkimjp.konpy2.util :refer [now user]]
-   [hkimjp.konpy2.restrictions :as r]))
+   [hkimjp.konpy2.restrictions :as r]
+   [hkimjp.konpy2.util :refer [now user]]))
 
 (defn comment!
   "send comments to `e`.
@@ -16,6 +16,8 @@
            :data  {:to to :author author :comment comment :pid pid :pt pt}})
   (let [author (user request)]
     (try
+      (when-not (re-find #"\S" comment)
+        (throw (Exception. "empty comment")))
       (r/before-comment author)
       (ds/put! {:comment/status "yes"
                 :author author
@@ -34,5 +36,7 @@
 
 (defn hx-comment [{{:keys [e]} :path-params}]
   (t/log! {:level :info :id "hx-comment"})
-  (hx [:div (:comment (ds/pl (parse-long e)))]))
+  (hx [:div (-> (:comment (ds/pl (parse-long e)))
+                md/parse
+                md/->hiccup)]))